Bugfix: fix a theoretical race condition.
If thread a uses dpa_u_bo_intern, and then thread a drops the reference,
but thread b then aquires the same entry using dpa_u_bo_intern and
drops the reference again and frees the entry using the dpa__u_bo_unique_hashmap_destroy
before the dpa__u_bo_unique_hashmap_destroy function in thread a can lock and free the entry,
it would try to free it again.

That has now been fixed, by incrementing the refcount if we find the entry and it's refcount was 0 once more.
We then check in dpa__u_bo_unique_hashmap_destroy after aquiring the lock if the refcount is 0,
if not, re release the lock and drop that reference.