Replace all the buffer object structures.
This rewrite, just like the stuff it replaced, is still incomplete and a work in progress.
Instead of putting a type in the buffer objects, we now use tagged pointers, or rather just 8 bytes aligned like a uint64_t.
Inline BOs are now at most 7 characters long, and the type is now a bitfield, stored in the first byte.
The BO types are now named differently. Currently, only the simple dpa_u_bo_t and dpa_u_bo_ro_t are exposed to be used directly.
Internally, dpa_u_bo_t is always used. Turning them into a dpa_a_*_t tagged pointer always creates a new object on the stack. This is to avaid potential aliasing issues.
The type of a tagged bo pointer can not be changed in the function it was passed to (by value). This may introduce the need of an additional indirection.